исправлена задача 6

+++ Zadacha6.java
@@ -11,11 +11,10 @@ public class Zadacha6 {
         int Y = sc.nextInt();
         int Z = sc.nextInt();
 
-
-        int A = X+Y+Z;
-        int S = A/3;
+        double A = X+Y+Z;
+        double S = A/3;
         System.out.println("среднее арифметическое чисел X,Y,Z: "+S);
-        int D = S/2;
+        int D = (int)S/2;
         System.out.println("Деление среднего арифметического на 2 без остатка: "+D);
         //Если полученное число больше 3, то вывести на экран сообщение "Программа выполнена корректно"
         if (D > 3);
